+++ daemons/mrpd/mvrp.c
@@ -880,7 +880,7 @@ mvrp_emit_vidvectors(unsigned char *msgbuf, unsigned char *msgbuf_eof,
 	*mrpdu_msg_ptr = 0;
 	mrpdu_msg_ptr++;
 
-	*bytes_used = (mrpdu_msg_ptr - msgbuf);
+	*bytes_used = (int)(mrpdu_msg_ptr - msgbuf);
 	return 0;
  oops:
 	/* an internal error - caller should assume TXLAF */
@@ -987,9 +987,9 @@ int mvrp_txpdu(void)
 	} else
 		goto out;
 
-	msgbuf_len = mrpdu_msg_ptr - msgbuf;
+	msgbuf_len = (int)(mrpdu_msg_ptr - msgbuf);
 
-	bytes = mrpd_send(mvrp_socket, msgbuf, msgbuf_len, 0);
+	bytes = (int)mrpd_send(mvrp_socket, msgbuf, msgbuf_len, 0);
 #if LOG_MVRP
 	mrpd_log_printf("MVRP send PDU\n");
 #endif
